need debug memory buffer fill according to pragma

+++ src/loader_utils/n_atof.c
@@ -1,5 +1,5 @@
 #include "object_loader.h"
-
+#include <math.h>
 /*
 ** Convert string of size n into float
 */
@@ -9,7 +9,7 @@ float n_atof(const char* str, size_t size)
   float  ret;
 	float  negate;
 	int 	 point_seen;
-	size_t i;
+	int i;
 	int		 digit;
 
 	ret = 0;
@@ -22,7 +22,7 @@ float n_atof(const char* str, size_t size)
 
 	i = 0;
 	point_seen = -1;
-	while (i < size && str[i]) {
+	while (i < (int) size && str[i]) {
 		if (str[i] == '.'){
 			point_seen = i;
 		} else {
@@ -30,8 +30,7 @@ float n_atof(const char* str, size_t size)
 
 			if (digit >= 0 && digit <= 9) {
 				if (point_seen > -1) {
-
-					ret +=  powf(0.1f, i - point_seen) * (float)digit;
+					ret +=  powf(0.1f, (float)(i - point_seen)) * (float)digit;
 				} else {
 					ret = ret * 10.0f + (float)digit;
 				}
